Move more files into src/webhooks (#36578)
